auth/pyproject.toml
[tool.poetry]
name = "authservice"
version = "0.1.0"
description = ""
authors = ["Кирилл Петухов <test@mail.ru>", "Калганов Данила"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "2.0.22"
fastapi = "0.100.1"
redis = "5.0.1"
pydantic = "2.1.1"
werkzeug = "3.0.0"
alembic = "^1.12.0"
uvicorn = "^0.23.2"
asyncpg = "^0.28.0"
greenlet = "^3.0.0"
pydantic-settings = "2.0.2"
orjson = "^3.9.9"
async-fastapi-jwt-auth = "0.6.1"
pytest = "6.2.5"
pytest-asyncio = "0.19.0"
slowapi = "0.1.8"
authlib = "1.2.1"
itsdangerous = "2.1.2"
httpx = "0.25.1"
opentelemetry-api = "1.21.0"
opentelemetry-exporter-jaeger = "1.21.0"
opentelemetry-exporter-otlp = "1.21.0"
opentelemetry-instrumentation-fastapi = "0.42b0"
opentelemetry-sdk = "1.21.0"


[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
